permission updated

+++ app/Http/Controllers/Backend/ModuleController.php
@@ -4,8 +4,9 @@ namespace App\Http\Controllers\Backend;
 
 use App\Models\Module;
 use Illuminate\Support\Str;
-use Brian2694\Toastr\Facades\Toastr;
 use App\Http\Controllers\Controller;
+use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\ModuleStoreRequest;
 
 class ModuleController extends Controller
@@ -13,6 +14,8 @@ class ModuleController extends Controller
 
     public function index()
     {
+        Gate::authorize('index-module');
+
         $modules=Module::select(['id','module_name','module_slug','updated_at'])->latest()->get();
         return view('pages.module.index',compact('modules'));
     }
@@ -20,12 +23,16 @@ class ModuleController extends Controller
 
     public function create()
     {
+        Gate::authorize('create-module');
+
         return view('pages.module.create');
     }
 
 
     public function store(ModuleStoreRequest $request)
     {
+        Gate::authorize('create-module');
+
         Module::updateOrCreate([
             'module_name' => $request->module_name,
             'module_slug' => Str::slug($request->module_name),
@@ -44,6 +51,8 @@ class ModuleController extends Controller
 
     public function edit(string $id)
     {
+        Gate::authorize('edit-module');
+
         $module=Module::find($id);
         return view('pages.module.edit',compact('module'));
     }
@@ -51,6 +60,8 @@ class ModuleController extends Controller
 
     public function update(ModuleStoreRequest $request, string $id)
     {
+        Gate::authorize('edit-module');
+
         $module = Module::find($id);
         $module->update([
             'module_name' => $request->module_name,
@@ -64,6 +75,8 @@ class ModuleController extends Controller
 
     public function destroy(string $id)
     {
+        Gate::authorize('delete-module');
+
         $module = Module::find($id);
         $module->delete();
