+++ app/Http/Controllers/Backend/PermissionController.php
@@ -7,6 +7,7 @@ use App\Models\Permission;
 use Illuminate\Support\Str;
 use App\Http\Controllers\Controller;
 use Brian2694\Toastr\Facades\Toastr;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\PermissionStoreRequest;
 
 class PermissionController extends Controller
@@ -14,6 +15,8 @@ class PermissionController extends Controller
 
     public function index()
     {
+        Gate::authorize('index-permission');
+
         $permissions=Permission::with(['module:id,module_name'])->select(['id','module_id','permission_name','permission_slug','updated_at'])->latest()->get();
 
         return view('pages.permission.index',compact('permissions'));
@@ -22,6 +25,8 @@ class PermissionController extends Controller
 
     public function create()
     {
+        Gate::authorize('create-permission');
+
         $modules = Module::select(['id', 'module_name'])->latest()->get();
 
         return view('pages.permission.create',compact('modules'));
@@ -30,6 +35,8 @@ class PermissionController extends Controller
 
     public function store(PermissionStoreRequest $request)
     {
+        Gate::authorize('create-permission');
+
         Permission::updateOrCreate([
             'module_id' => $request->module_id,
             'permission_name' => $request->permission_name,
@@ -49,6 +56,8 @@ class PermissionController extends Controller
 
     public function edit(string $id)
     {
+        Gate::authorize('edit-permission');
+
         $modules = Module::select(['id', 'module_name'])->latest()->get();
 
         $permission=Permission::find($id);
@@ -58,6 +67,8 @@ class PermissionController extends Controller
 
     public function update(PermissionStoreRequest $request, string $id)
     {
+        Gate::authorize('edit-permission');
+
         $permission = Permission::find($id);
         $permission->update([
             'module_id' => $request->module_id,
@@ -72,6 +83,8 @@ class PermissionController extends Controller
 
     public function destroy(string $id)
     {
+        Gate::authorize('delete-permission');
+
         $permission = Permission::find($id);
         $permission->delete();
